+++ aztk/plugins/__init__.py
@@ -1,4 +1,4 @@
-from .plugin_models import *
+from .plugin_definition import *
 
 """
 This module contains models and utils to be used in a plugin definition
